--- app/(dashboard)/(routes)/settings/_components/advanced-options/advanced-options-form.tsx
@@ -1,85 +1,0 @@
-'use client';
-
-import { zodResolver } from '@hookform/resolvers/zod';
-import { User } from '@prisma/client';
-import { useRouter } from 'next/navigation';
-import { useForm } from 'react-hook-form';
-import toast from 'react-hot-toast';
-import * as z from 'zod';
-
-import { Switch } from '@/components/ui';
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-} from '@/components/ui/form';
-import { fetcher } from '@/lib/fetcher';
-
-import { OtpForm } from './otp-form';
-
-type AdvancedOptionsFormProps = {
-  initialData: User;
-};
-
-const formSchema = z.object({
-  isPublic: z.boolean().default(false),
-});
-
-export const AdvancedOptionsForm = ({ initialData }: AdvancedOptionsFormProps) => {
-  const router = useRouter();
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      isPublic: Boolean(initialData.isPublic),
-    },
-  });
-
-  const { isSubmitting, isValid } = form.formState;
-
-  const handleSubmit = async (values: z.infer<typeof formSchema>) => {
-    try {
-      await fetcher.patch(`/api/users/${initialData.id}`, { body: values });
-      toast.success('Visibility updated');
-      router.refresh();
-    } catch (error) {
-      toast.error('Something went wrong!');
-    }
-  };
-
-  return (
-    <div className="flex flex-col gap-4 mt-8">
-      <p className="font-medium text-xl">Advanced Options</p>
-      <OtpForm initialData={initialData} />
-      <Form {...form}>
-        <form onSubmit={form.handleSubmit(handleSubmit)}>
-          <FormField
-            control={form.control}
-            name="isPublic"
-            render={({ field }) => (
-              <FormItem className="flex flex-row items-center justify-between space-x-3 space-y-0 rounded-md border p-4">
-                <div className="space-y-0.5">
-                  <FormLabel>Enable Public Profile</FormLabel>
-                  <FormDescription className="text-xs">
-                    Change the visibility of your profile
-                  </FormDescription>
-                </div>
-                <FormControl>
-                  <Switch
-                    aria-readonly
-                    checked={field.value}
-                    disabled={!isValid || isSubmitting}
-                    onCheckedChange={field.onChange}
-                    type="submit"
-                  />
-                </FormControl>
-              </FormItem>
-            )}
-          />
-        </form>
-      </Form>
-    </div>
-  );
-};
