+++ app/(dashboard)/(routes)/settings/page.tsx
@@ -5,7 +5,7 @@ import { DeleteAccountModal } from '@/components/modals/delete-account-modal';
 import { Button } from '@/components/ui';
 import { db } from '@/lib/db';
 
-import { AdvancedOptionsForm } from './_components/advanced-options/advanced-options-form';
+import { AdvancedOptions } from './_components/advanced-options/advanced-options';
 import { GeneralSettingsForm } from './_components/general-settings-form';
 
 const SettingsPage = async () => {
@@ -17,7 +17,7 @@ const SettingsPage = async () => {
       <h1 className="text-2xl font-medium">General Settings</h1>
       <div className="mt-12">
         <GeneralSettingsForm initialData={userInfo!} />
-        <AdvancedOptionsForm initialData={userInfo!} />
+        <AdvancedOptions initialData={userInfo!} />
       </div>
       <DeleteAccountModal userId={user?.userId} email={user?.email}>
         <div className="flex items-center gap-x-2 mt-8">
